% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olhovivo.R
\name{draw_bus}
\alias{draw_bus}
\title{Desenha ônibus em tempo real}
\usage{
draw_bus(.data, map = NULL)
}
\arguments{
\item{.data}{data frame retornado pela função \code{collect_bus}.}

\item{map}{Mapa para adicionar layes, caso já exista.}
}
\description{
A partir de dados coletados da API OlhoVivo, desenha os ônibus no mapa
gerado pelo pacote leaflet.
}
\examples{
\dontrun{
trip_ids <- search_path(
  end1 = 'Avenida 9 de Julho, 2000, São Paulo', 
  end2 = 'Av. Pres. Juscelino Kubitschek, 500, São Paulo'
)

stop_ids <- nearby_stops('Avenida Paulista, 1079, São Paulo')

trip_ids \%>\% 
  collect_bus(trip_id, 'trip') \%>\% 
  draw_bus()
}
}
