% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{draw_stops}
\alias{draw_stops}
\title{Desenha os pontos de ônibus}
\usage{
draw_stops(.data)
}
\arguments{
\item{.data}{dados}
}
\value{
Um objeto HTML widget do pacote \code{leaflet}.
}
\description{
A partir de um objeto retornado pela função \code{nearby_stops}, desenha
um mapa com o lugar escolhido, o raio selecionado e os pontos de ônibus
próximos.
}
\examples{
\dontrun{
nearby_stops('Avenida Paulista, 1079') \%>\% draw_stops()
}
}
