% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sptrans.R
\name{leaflet_desenha_shapes}
\alias{leaflet_desenha_shapes}
\title{Desenha as shapes das linhas de onibus no mapa nao usa API}
\usage{
leaflet_desenha_shapes(cod_linhas = NULL, shape_ids = NULL, m = NULL)
}
\description{
Desenha as shapes das linhas de onibus no mapa nao usa API
}

