% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gtfs.R
\name{nearby_stops}
\alias{nearby_stops}
\title{Identifica pontos de ônibus próximos}
\usage{
nearby_stops(end = NULL, radius = 300, lon = NULL, lat = NULL)

nearby_stops(end = NULL, radius = 300, lon = NULL, lat = NULL)
}
\arguments{
\item{end}{Endereço para procurar (como se estivesse pesquisando no google maps).}

\item{radius}{Raio em metros para procura dos pontos mais próximos. Por
padrão vale 300 metros.}

\item{lon}{Longitude. Informar se o endereço for nulo.}

\item{lat}{Latitude. Informar se o endereço for nulo.}

\item{end}{Endereço para procurar (como se estivesse pesquisando no google maps).}

\item{radius}{Raio em metros para procura dos pontos mais próximos. Por
padrão vale 300 metros.}

\item{lon}{Longitude. Informar se o endereço for nulo.}

\item{lat}{Latitude. Informar se o endereço for nulo.}
}
\value{
Um data frame contendo um subset de \code{stops}
  (pacote \code{spgtfs}) com os pontos mais próximos, ordenados
  por proximidade

Um data frame contendo um subset de \code{stops}
  (pacote \code{spgtfs}) com os pontos mais próximos, ordenados
  por proximidade
}
\description{
Calcula distâncias entre um ponto escolhido e os pontos de ônibus,
retornando somente aqueles que estão dentro do raio informado.

Calcula distâncias entre um ponto escolhido e os pontos de ônibus,
retornando somente aqueles que estão dentro do raio informado.
}
\examples{
nearby_stops('Avenida Paulista, 1079')
nearby_stops('Avenida Paulista, 1079', radius = 200)
nearby_stops(lon = -46.6527, lat = -23.5648)
nearby_stops('Avenida Paulista, 1079')
nearby_stops('Avenida Paulista, 1079', radius = 200)
nearby_stops(lon = -46.6527, lat = -23.5648)
}

