% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{search_path}
\alias{search_path}
\title{Procura trajetos de ônibus que passam pelos pontos.}
\usage{
search_path(end1 = NULL, end2 = NULL, radius1 = 300, radius2 = 300,
  lon1 = NULL, lat1 = NULL, lon2 = NULL, lat2 = NULL)
}
\arguments{
\item{end1}{Endereço de partida.}

\item{end2}{Endereço de destino.}

\item{radius1}{Raio em metros para procura dos pontos mais próximos do ponto
de partida. Por padrão vale 500 metros.}

\item{radius2}{Raio em metros para procura dos pontos mais próximos do ponto
de destino Por padrão vale 500 metros.}

\item{lon1}{Longitude de partida. Informar se o endereço for nulo.}

\item{lat1}{Latitude de partida. Informar se o endereço for nulo.}

\item{lon2}{Longitude de destino. Informar se o endereço for nulo.}

\item{lat2}{Latitude de destino. Informar se o endereço for nulo.}
}
\value{
Um data frame contendo um subset de \code{trips} 
  (pacote \code{spgtfs}) com as linhas que atendem aos critérios.
}
\description{
Obtém os pontos de ônibus mais próximos de cada ponto informado e procura
por linhas que passam por esses pontos.
}
\examples{
\dontrun{
search_path(from = 'Avenida 9 de Julho, 2000', 
            to = 'Av. Pres. Juscelino Kubitschek, 500')
}
}
