% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{draw_paths}
\alias{draw_paths}
\title{Desenha as trips}
\usage{
draw_paths(.data)
}
\arguments{
\item{.data}{dados}
}
\value{
Um objeto HTML widget do pacote \code{leaflet}.
}
\description{
A partir de um objeto retornado pela função \code{search_paths}, desenha
um mapa com os lugares escolhidos, os raios selecionados, os pontos de 
ônibus próximos e os caminhos.
}
\examples{
\dontrun{
search_path(end1 = 'Avenida 9 de Julho, 2000, São Paulo', 
            end2 = 'Av. Pres. Juscelino Kubitschek, 500, São Paulo') \%>\%
 draw_paths()
}
}
